+++ src/lib/__init__.py
@@ -14,5 +14,5 @@ from .bodefano import BodeFano
 from .stabcircle import StabilityCircle
 from .shortstr import shorten_string_list
 from .clipboard import Clipboard
-from .settings import SParamViewerAppSettings, Settings, PlotType, SmithNorm, TdResponse, YQuantity, PhaseProcessing, PhaseUnit, CsvSeparator, CursorSnap, ColorAssignment, Parameters, LogNegativeHandling, MainWindowLayout, LargeMatrixBehavior
+from .settings import SParamViewerAppSettings, Settings, PlotType, SmithNorm, TdResponse, YQuantity, PhaseProcessing, PhaseUnit, CsvSeparator, CursorSnap, ColorAssignment, Parameters, LogNegativeHandling, MainWindowLayout, LargeMatrixBehavior, GuiColorScheme
 from .lock import Lock